wooden_calculator_wastage.java
there is a company who manufacture wooden blinds. the number of  strips required to make a blind is depend on height of window. This strip is cut from wooden piece. these wooden pieces are comes in three different size 4 feet, 5 feet and 6 feet.  WAP code in java to find the which wooden piece size will be best to make strips and will give minimum wastage. take the size of strip from user and number of strip from user.



taste case: - 
size- 30 inch
number is 10

1) if 4 feet is used = 48inch
one wooden piece = 48 inch
48 - 30 = 8 inch 

now 8 inch is the wastage which can't be used because it is less then 30. 
so total wastage is 8 * 10. inch 
total pieces required are 10.

2) if take 5 feet = 60 inch

60 - 30 = 30 
now 30 is wastage but wastage and strip size is equal so it can be used as strip. 
means in 1 wooden piece, we can make two strips.
so total wooden piece  required  to make strips =  10 / 2 = 5 wooden piece .

wastage = 0.

3) if using 72

72 - 30 = 42 
now again 42 will be cut off size 30 beacuse its is more then strip size.
so, in 1 wooden piece 2 strip made and wastage is 12 inch.

total wooden strips required  = 10 / 2 = 5.
wastage = 12 * 5 = 60 inch.

so best wooden piece will be 5 feet.